Implement fix for issue #1
On some platforms, dump1090 is sending a timestamp with more resolution than python can handle. In issue #1, the timestamps have the format: `07:37:23.1568119043`. Python datetime's `%f` directive supports from one to six digits, and was causing a `ValueError` for the extra digits. Adding a restriction `[0:15]` for the time string simply drops the additional digits from the end of the string, preventing an error for these situations.